Handle linux and windows platform config dirs
	modified:   lib/torrentprocessor.rb
	modified:   lib/torrentprocessor/tpsetup.rb
	modified:   lib/torrentprocessor/utility/loggers.rb

+++ lib/torrentprocessor/utility/loggers.rb
@@ -56,24 +56,31 @@ class FileLogger
     rotate_log
     timestamp = DateTime.now.strftime()
 
-    File.open(logpath, 'a') { |f| f.write( "#{timestamp}: #{msg}\n"); f.flush; }
+    if ! File.exist? logpath
+      FileUtils.mkdir_p logdir
+      File.open(logpath, 'w') { |f| f.write( "#{timestamp}: #{msg}\n"); f.flush; }
+    else
+      File.open(logpath, 'a') { |f| f.write( "#{timestamp}: #{msg}\n"); f.flush; }
+    end
   end
 
   def self.rotate_log
     return if max_log_size == 0
 
-    if FileTest.size(logpath) > max_log_size
+    if File.exist? logpath
+      if FileTest.size(logpath) > max_log_size
 
-      # Used to determine if a file was still open when receiving EACCESS errors.
-      #puts "#"*20
-      #ObjectSpace.each_object(File) { |f| p f if f.path.include?(logpath) && !f.closed? }
-      #puts "#"*20
+        # Used to determine if a file was still open when receiving EACCESS errors.
+        #puts "#"*20
+        #ObjectSpace.each_object(File) { |f| p f if f.path.include?(logpath) && !f.closed? }
+        #puts "#"*20
 
-      FileUtils.rm(logpath + '.3') if File.exists?(logpath + '.3')
-      FileUtils.mv(logpath + '.2', logpath + '.3') if File.exists?(logpath + '.2')
-      FileUtils.mv(logpath + '.1', logpath + '.2') if File.exists?(logpath + '.1')
-      FileUtils.mv(logpath, logpath + '.1') if File.exists?(logpath)
-    end # if
+        FileUtils.rm(logpath + '.3') if File.exists?(logpath + '.3')
+        FileUtils.mv(logpath + '.2', logpath + '.3') if File.exists?(logpath + '.2')
+        FileUtils.mv(logpath + '.1', logpath + '.2') if File.exists?(logpath + '.1')
+        FileUtils.mv(logpath, logpath + '.1') if File.exists?(logpath)
+      end # if
+    end
   end
 end
